update scenario types
