+++ frontend/utils/validation.ts
@@ -10,34 +10,5 @@ export { Hero, TankHero, DamageHero, SupportHero };
 import { Map, ControlMap, EscortMap, FlashpointMap, HybridMap, PushMap, ClashMap } from '../types/Maps';
 export { Map, ControlMap, EscortMap, FlashpointMap, HybridMap, PushMap, ClashMap };
 
-export const Scenario = z.object({
-    map: Map,
-    point: z.coerce.number().int().gte(0).lte(2),
-    teams: z.array(z.object({
-        players: z.array(z.object({
-            hero: Hero,
-            ultCharge: z.number().int().gte(0).lte(100),
-            position: z.object({
-                x: z.number().int(),
-                y: z.number().int(),
-                z: z.number().int(),
-            }),
-        })),
-    })),
-});
-
-export const ScenarioString = z.string().transform((str) => {
-    try {
-        return JSON.parse(str);
-    } catch (e) {
-        throw new Error('Invalid JSON');
-    }
-}).pipe(Scenario).transform((scenario) => JSON.stringify(scenario));
-
-export const ScenarioMetadata = z.object({
-    name: z.string().min(1).max(255),
-    ownerID: z.number(),
-    createdAt: z.string().datetime().transform((str) => new Date(str)),
-    lastEdited: z.string().datetime().transform((str) => new Date(str)),
-    scenarioData: ScenarioString,
-});
+import { Player, Team, Scenario, ScenarioString, ScenarioMetadata } from '../types/Scenario';
+export { Player, Team, Scenario, ScenarioString, ScenarioMetadata };
